Created the new folder
This is in a discrete commit because Github is silly and does not have basic management tools. Thus, I am first forced to create a useless folder before I can actually put the new files in. No, you do not need a separate command line terminal program to manage files.